replaced export with backpressure handling export
