+++ src/api.ts
@@ -1,5 +1,4 @@
 export { dev } from './api/dev';
 export { build } from './api/build';
-// export { export } from './api/export';
-export { export } from './api/export_backpressure';
+export { export } from './api/export';
 export { find_page } from './api/find_page';
